+++ components/select/constants.js
@@ -1,0 +1,12 @@
+export const colors = {
+  NEUTRAL_LIGHTEST: '#fff',
+  NEUTRAL_LIGHT: '#f7f7fa',
+  NEUTRAL: '#e4e4e6',
+  NEUTRAL_DARK: '#c0c0c4',
+  NEUTRAL_DARKEST: '#82828c',
+  TEAL_LIGHTEST: '#e1ecfa',
+  TEAL_LIGHT: '#c1cede',
+  TEAL: '#64788f',
+  TEAL_DARK: '#344b63',
+  TEAL_DARKEST: '#2a3b4d',
+};